Added On_Attack Effect

+++ src/Battle/BattleReport.java
@@ -3,7 +3,7 @@ package Battle;
 import java.util.HashMap;
 
 public class BattleReport {
-    public HashMap<ReportIndex,Float> report;
+    public HashMap<ReportIndex,Integer> report;
 
     public BattleReport(){
         report = new HashMap<>();
